Refactor cloud function into multiple files part 4 - peerings
