+++ examples/cloud-operations/network-dashboard/cloud-function/metrics/ilb_fwrules.py
@@ -1,8 +1,7 @@
 from code import interact
 from collections import defaultdict
 from google.protobuf import field_mask_pb2
-from google.cloud import asset_v1
-from . import metrics, networks
+from . import metrics, networks, limits
 import main as main
 
 def get_forwarding_rules_dict(config, layer: str):
@@ -62,8 +61,7 @@ def get_forwarding_rules_data(config, metrics_dict, forwarding_rules_dict,
   for project in config["monitored_projects"]:
     network_dict = networks.get_networks(config, project)
 
-    #TODO Workaround
-    current_quota_limit = main.get_quota_current_limit(
+    current_quota_limit = limits.get_quota_current_limit(
         f"projects/{project}", config["limit_names"][layer])
 
     if current_quota_limit is None:
@@ -76,8 +74,7 @@ def get_forwarding_rules_data(config, metrics_dict, forwarding_rules_dict,
     current_quota_limit_view = main.customize_quota_view(current_quota_limit)
 
     for net in network_dict:
-      #TODO Workaround
-      main.set_limits(net, current_quota_limit_view, limit_dict)
+      limits.set_limits(net, current_quota_limit_view, limit_dict)
 
       usage = 0
       if net['self_link'] in forwarding_rules_dict:
