+++ examples/cloud-operations/network-dashboard/cloud-function/metrics/limits.py
@@ -1,3 +1,8 @@
+import main
+from google.api_core import exceptions
+from google.cloud import monitoring_v3
+from . import metrics
+
 def get_ppg(network_link, limit_dict):
   '''
     Checks if this network has a specific limit for a metric, if so, returns that limit, if not, returns the default limit.
@@ -16,3 +21,133 @@ def get_ppg(network_link, limit_dict):
     else:
       print(f"Error: limit not found for {network_link}")
       return 0
+
+def set_limits(network_dict, quota_limit, limit_dict):
+  '''
+    Updates the network dictionary with quota limit values.
+
+      Parameters:
+        network_dict (dictionary of string: string): Contains network information.
+        quota_limit (list of dictionaries of string: string): Current quota limit.
+        limit_dict (dictionary of string:int): Dictionary with the network link as key and the limit as value
+      Returns:
+        None
+  '''
+
+  network_dict['limit'] = None
+
+  if quota_limit:
+    for net in quota_limit:
+      if net['network_id'] == network_dict['network_id']:
+        network_dict['limit'] = net['value']
+        return
+
+  network_link = f"https://www.googleapis.com/compute/v1/projects/{network_dict['project_id']}/global/networks/{network_dict['network_name']}"
+
+  if network_link in limit_dict:
+    network_dict['limit'] = limit_dict[network_link]
+  else:
+    if 'default_value' in limit_dict:
+      network_dict['limit'] = limit_dict['default_value']
+    else:
+      print(f"Error: Couldn't find limit for {network_link}")
+      network_dict['limit'] = 0
+
+def get_quota_current_limit(project_link, metric_name):
+  '''
+    Retrieves limit for a specific metric.
+
+      Parameters:
+        project_link (string): Project link.
+        metric_name (string): Name of the metric.
+      Returns:
+        results_list (list of string): Current limit.
+  '''
+  client, interval = main.create_client()
+
+  try:
+    results = client.list_time_series(
+        request={
+            "name": project_link,
+            "filter": f'metric.type = "{metric_name}"',
+            "interval": interval,
+            "view": monitoring_v3.ListTimeSeriesRequest.TimeSeriesView.FULL
+        })
+    results_list = list(results)
+    return results_list
+  except exceptions.PermissionDenied as err:
+    print(
+        f"Warning: error reading quotas for {project_link}. " +
+        f"This can happen if you don't have permissions on the project, for example if the project is in another organization or a Google managed project"
+    )
+  return None
+
+
+def count_effective_limit(config, project_id, network_dict, usage_metric_name,
+                          limit_metric_name, utilization_metric_name,
+                          limit_dict):
+  '''
+    Calculates the effective limits (using algorithm in the link below) for peering groups and writes data (usage, limit, utilization) to the custom metrics.
+    Source: https://cloud.google.com/vpc/docs/quota#vpc-peering-effective-limit
+
+      Parameters:
+        project_id (string): Project ID for the project to be analyzed.
+        network_dict (dictionary of string: string): Contains all required information about the network to get the usage, limit and utilization.
+        usage_metric_name (string): Name of the custom metric to be populated for usage per VPC peering group.
+        limit_metric_name (string): Name of the custom metric to be populated for limit per VPC peering group.
+        utilization_metric_name (string): Name of the custom metric to be populated for utilization per VPC peering group.
+        limit_dict (dictionary of string:int): Dictionary containing the limit per peering group (either VPC specific or default limit).
+      Returns:
+        None
+  '''
+
+  if network_dict['peerings'] == []:
+    return
+
+  # Get usage: Sums usage for current network + all peered networks
+  peering_group_usage = network_dict['usage']
+  for peered_network in network_dict['peerings']:
+    if 'usage' not in peered_network:
+      print(
+          f"Can not add metrics for peered network in projects/{project_id} as no usage metrics exist due to missing permissions"
+      )
+      continue
+    peering_group_usage += peered_network['usage']
+
+  network_link = f"https://www.googleapis.com/compute/v1/projects/{project_id}/global/networks/{network_dict['network_name']}"
+
+  # Calculates effective limit: Step 1: max(per network limit, per network_peering_group limit)
+  limit_step1 = max(network_dict['limit'],
+                    get_ppg(network_link, limit_dict))
+
+  # Calculates effective limit: Step 2: List of max(per network limit, per network_peering_group limit) for each peered network
+  limit_step2 = []
+  for peered_network in network_dict['peerings']:
+    peered_network_link = f"https://www.googleapis.com/compute/v1/projects/{peered_network['project_id']}/global/networks/{peered_network['network_name']}"
+
+    if 'limit' in peered_network:
+      limit_step2.append(
+          max(peered_network['limit'],
+              get_ppg(peered_network_link, limit_dict)))
+    else:
+      print(
+          f"Ignoring projects/{peered_network['project_id']} for limits in peering group of project {project_id} as no limits are available."
+          +
+          "This can happen if you don't have permissions on the project, for example if the project is in another organization or a Google managed project"
+      )
+
+  # Calculates effective limit: Step 3: Find minimum from the list created by Step 2
+  limit_step3 = 0
+  if len(limit_step2) > 0:
+    limit_step3 = min(limit_step2)
+
+  # Calculates effective limit: Step 4: Find maximum from step 1 and step 3
+  effective_limit = max(limit_step1, limit_step3)
+  utilization = peering_group_usage / effective_limit
+
+  metrics.write_data_to_metric(config, project_id, peering_group_usage, usage_metric_name,
+                       network_dict['network_name'])
+  metrics.write_data_to_metric(config, project_id, effective_limit, limit_metric_name,
+                       network_dict['network_name'])
+  metrics.write_data_to_metric(config, project_id, utilization, utilization_metric_name,
+                       network_dict['network_name'])
